+++ src/core/embedding_model/mod.rs
@@ -7,6 +7,7 @@ pub mod request;
 
 use crate::error::Result;
 use async_trait::async_trait;
+use std::collections::HashMap;
 
 use derive_builder::Builder;
 #[cfg(feature = "embedding-model-request")]
@@ -22,6 +23,8 @@ pub struct EmbeddingModelOptions {
     /// model used and if the provider returns an error for invalid dimensions `embed`
     /// will propagate the error.
     pub dimensions: Option<usize>,
+    /// Custom HTTP headers to include in the request.
+    pub headers: Option<HashMap<String, String>>,
 }
 
 impl EmbeddingModelOptions {